pyproject.toml
[project]
name = "lab7"
version = "0.1.0"
description = ""
authors = [
    {name = "ech0p1ng",email = "alexey2003petrov@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "mypy (>=1.19.0,<2.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg (>=3.3.2,<4.0.0)",
    "logger (>=1.4,<2.0)",
    "fastapi[standard] (>=0.127.0,<0.128.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "minio (>=7.2.20,<8.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "types-python-jose (>=3.5.0.20250531,<4.0.0.0)",
    "types-passlib (>=1.7.7.20250602,<2.0.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "numpy (>=2.4.0,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "aiohttp (>=3.13.2,<4.0.0)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "asyncio (>=4.0.0,<5.0.0)",
    "debugpy (>=1.8.19,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
